docs(project): add readme.md
This commit adds a `README.md` file.. The `README.md` file provides a
comprehensive overview of the modpack, including a list of mods and
libraries used in the modpack.

The `README.md` file also provides usage instructions for each mod and
library, and specifies whether they need to be installed on the client
side, server side, or both.