refactor(structure): separate mods into client and server
This commit restructures the modpack to separate mods into client and
server directories. This allows for faster and more convenient mod
installation. The new structure is as follows:

technomagic-modpack-minecraft-forge-1.19.2
├──client
│   ├──libs
│   │   └──lib1.jar
│   │   └──lib2.jar...
│   └──mods
│       └──mod1.jar
│       └──mod2.jar...
├──server
│   ├──libs
│   │   └──lib1.jar
│   │   └──lib2.jar...
│   └──mods
│       └──mod1.jar
│       └──mod2.jar...
└──README.md

This change will make it easier for users to install the mods they need
depending on whether they are playing on the client or the server.